[FEAT] 응원톡 관리 화면에서의 응원톡 전체 조회

+++ src/main/java/com/sports/server/command/leagueteam/application/LeagueTeamService.java
@@ -1,6 +1,5 @@
 package com.sports.server.command.leagueteam.application;
 
-import com.sports.server.auth.exception.AuthorizationErrorMessages;
 import com.sports.server.command.league.domain.League;
 import com.sports.server.command.leagueteam.domain.LeagueTeam;
 import com.sports.server.command.leagueteam.domain.LeagueTeamPlayer;
@@ -10,8 +9,8 @@ import com.sports.server.command.leagueteam.dto.LeagueTeamPlayerRequest;
 import com.sports.server.command.leagueteam.dto.LeagueTeamRequest;
 import com.sports.server.command.member.domain.Member;
 import com.sports.server.common.application.EntityUtils;
+import com.sports.server.common.application.PermissionValidator;
 import com.sports.server.common.exception.NotFoundException;
-import com.sports.server.common.exception.UnauthorizedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -33,7 +32,8 @@ public class LeagueTeamService {
     private final EntityUtils entityUtils;
 
     public void register(final Long leagueId, final Member manager, final LeagueTeamRequest.Register request) {
-        League league = getLeagueAndCheckPermission(leagueId, manager);
+        League league = entityUtils.getEntity(leagueId, League.class);
+        PermissionValidator.checkPermission(league, manager);
 
         String imgUrl = changeLogoImageUrlToBeSaved(request.logoImageUrl());
         LeagueTeam leagueTeam = request.toEntity(manager, league, imgUrl);
@@ -46,7 +46,9 @@ public class LeagueTeamService {
     }
 
     public void update(Long leagueId, LeagueTeamRequest.Update request, Member manager, Long teamId) {
-        getLeagueAndCheckPermission(leagueId, manager);
+        League league = entityUtils.getEntity(leagueId, League.class);
+        PermissionValidator.checkPermission(league, manager);
+
         LeagueTeam leagueTeam = getLeagueTeam(teamId);
 
         leagueTeam.updateInfo(request.name(), changeLogoImageUrlToBeSaved(request.logoImageUrl()));
@@ -57,7 +59,9 @@ public class LeagueTeamService {
     }
 
     public void delete(Long leagueId, Member manager, Long teamId) {
-        League league = getLeagueAndCheckPermission(leagueId, manager);
+        League league = entityUtils.getEntity(leagueId, League.class);
+        PermissionValidator.checkPermission(league, manager);
+
         LeagueTeam leagueTeam = entityUtils.getEntity(teamId, LeagueTeam.class);
         leagueTeam.isParticipate(league);
 
@@ -102,7 +106,8 @@ public class LeagueTeamService {
     }
 
     public void deleteLogoImage(Long leagueId, Member manager, Long teamId) {
-        getLeagueAndCheckPermission(leagueId, manager);
+        League league = entityUtils.getEntity(leagueId, League.class);
+        PermissionValidator.checkPermission(league, manager);
 
         LeagueTeam leagueTeam = entityUtils.getEntity(teamId, LeagueTeam.class);
         leagueTeam.deleteLogoImageUrl();
@@ -115,13 +120,4 @@ public class LeagueTeamService {
         return logoImageUrl.replace(originPrefix, replacePrefix);
     }
 
-    private League getLeagueAndCheckPermission(final Long leagueId, final Member manager) {
-        League league = entityUtils.getEntity(leagueId, League.class);
-
-        if (!league.isManagedBy(manager)) {
-            throw new UnauthorizedException(AuthorizationErrorMessages.PERMISSION_DENIED);
-        }
-
-        return league;
-    }
 }
